reworked autotranslate into translate
